+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Drop `colorize` which is GPL-licensed in favour of `rainbow`
 
 ## [0.15.0] - 2024-08-14
 - Fix Langchain::Assistant when llm is Anthropic
@@ -135,7 +136,7 @@
 - AWS Bedrock is available as an LLM provider. Available models from AI21, Cohere, AWS, and Anthropic.
 
 ## [0.7.3] - 2023-11-08
-- LLM response passes through the context in RAG cases 
+- LLM response passes through the context in RAG cases
 - Fix gpt-4 token length validation
 
 ## [0.7.2] - 2023-11-02
@@ -180,7 +181,7 @@
 
 ## [0.6.13] - 2023-08-23
 - Add `k:` parameter to all `ask()` vector search methods
-- Bump Faraday to 2.x 
+- Bump Faraday to 2.x
 
 ## [0.6.12] - 2023-08-13