+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+
+## [0.15.1] - 2024-08-19
 - Drop `to_bool` gem in favour of custom util class
 - Drop `colorize` which is GPL-licensed in favour of `rainbow`
 - Improve Langchain::Tool::Weather tool